+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Update Zcash and Tezos nodes.
+
 ## 4.3.0 (2024-04-22)
 
 - added: (Zcash) Orchard pool support